Added RecentActivity functionality, similar to the panel in the Rally tool. This required refactoring the class heirarchy to include DomainObjects as parents of Artifacts (since Conversation Posts are Domain Objects and not Artifacts).
I've pushed common functionality from Artifact into DomainObject and reworked the code to use DomainObjects.

The methods getLine1 and getLine2, which take DomainObject types, defined in RallyListActivity are now forcing us to cast the formal parameters in the implementations. We should probably refactor this to fix it. I've not figured out the best way to do this yet.